+++ test/controllers/home_controller_test.rb
@@ -40,6 +40,7 @@ class HomeControllerTest < ActionDispatch::IntegrationTest
     get "/"
 
     assert_response :success
+    assert_includes response.body, '<html data-theme="rounded">'
     assert_includes response.body, "Pages"
     assert_includes response.body, "Add page"
     assert_includes response.body, "Sign out"