+++ test/recording_studio_publishable_test.rb
@@ -21,10 +21,19 @@ class RecordingStudioPublishableTest < Minitest::Test
   def test_dummy_home_page_uses_default_layout_table
     view_path = File.expand_path("dummy/app/views/home/index.html.erb", __dir__)
     view_source = File.read(view_path)
+    layout_source = File.read(
+      File.expand_path("dummy/app/views/layouts/recording_studio/default_layout.html.erb", __dir__)
+    )
+    devise_layout = File.read(
+      File.expand_path("dummy/app/views/layouts/application.html.erb", __dir__)
+    )
 
     assert_includes view_source, "dummy_page_nav"
     assert_includes view_source, "FlatPack::Table::Component"
     assert_includes view_source, "Add page"
     refute_includes view_source, "Dummy publishables"
+    assert_includes layout_source, '<html data-theme="rounded">'
+    assert_includes layout_source, 'stylesheet_link_tag "flat_pack/application"'
+    assert_includes devise_layout, '<html data-theme="rounded">'
   end
 end